refactor: split backend modules and tests
