+++ src/backend/local/tests.rs
@@ -1,0 +1,15 @@
+use super::control::{mark_worker_stopping, StopTarget};
+use super::socket::check_no_existing_broker;
+use super::state::DEFAULT_WORKER_TTL_SECS;
+use super::*;
+use crate::backend::Backend;
+use crate::config::ResolvedConfig;
+use crate::protocol::STATUS_HISTORY_MAX;
+use std::path::{Path, PathBuf};
+use std::time::{SystemTime, UNIX_EPOCH};
+use tempfile::TempDir;
+use tokio::io::{AsyncBufReadExt, AsyncWriteExt, BufReader};
+
+include!("tests/part1.rs");
+include!("tests/part2.rs");
+include!("tests/part3.rs");
